Consolidate script plugins for publishing
